+++ source/background/library/messaging.js
@@ -9,8 +9,6 @@ import {
     changeVaultPassword,
     generateEntryPath,
     getArchive,
-    getEntry,
-    getFacades,
     getNameForSource,
     getSourceIDForVaultID,
     getSourcesInfo,
@@ -32,8 +30,9 @@ import { getConfig } from "../../shared/selectors/app.js";
 import { authenticateWithoutToken as authenticateGoogleDrive } from "./googleDrive.js";
 import { disableLoginsOnDomain, getDisabledDomains, removeDisabledFlagForDomain } from "./disabledLogin.js";
 import { getLogins, removeLogin, stopPromptForTab, updateLogin } from "./loginMemory.js";
-import { getSearch } from "./search.js";
+import { getCachedFacades, getSearch } from "./search.js";
 import { addAttachments, deleteAttachment, getAttachment, getAttachmentDetails } from "./attachments.js";
+import { writeToClipboard } from "./clipboard.js";
 
 export function clearSearchResults() {
     return getUnlockedSourcesCount().then(unlockedSources => {
@@ -119,6 +118,9 @@ function handleMessage(request, sender, sendResponse) {
         }
         case "clear-search":
             clearSearchResults();
+            return false;
+        case "copy-to-clipboard":
+            writeToClipboard(request.content);
             return false;
         case "create-vault-facade": {
             const { sourceID } = request;
@@ -416,7 +418,7 @@ function handleMessage(request, sender, sendResponse) {
 async function processSearchResults([entryResults, sources]) {
     const sourceIDs = {};
     const sourceNames = {};
-    const vaultFacades = await getFacades();
+    const vaultFacades = await getCachedFacades();
     const results = await Promise.all(
         entryResults.map(async entryResult => {
             const sourceID = (sourceIDs[entryResult.vaultID] =