+++ source/background/library/search.js
@@ -1,8 +1,17 @@
 import { Search } from "../../shared/library/buttercup.js";
 import BrowserStorageInterface, { getSyncStorage } from "./BrowserStorageInterface.js";
+import { getFacades } from "./facades.js";
 
 let __search = null,
+    __facades = [],
     __prepPromise = null;
+
+export async function getCachedFacades() {
+    if (__prepPromise) {
+        await __prepPromise;
+    }
+    return __facades;
+}
 
 export async function getSearch() {
     if (__prepPromise) {
@@ -18,5 +27,8 @@ export async function updateSearch(vaults) {
     }
     const storage = new BrowserStorageInterface(getSyncStorage());
     __search = new Search(vaults, storage);
-    __prepPromise = __search.prepare();
+    __prepPromise = __search.prepare().then(() => {
+        // Cache the facades
+        __facades = getFacades();
+    });
 }