+++ source/dialog/library/messaging.js
@@ -1,6 +1,10 @@
 import log from "../../shared/library/log.js";
 import i18n from "../../shared/i18n";
 
+export function copyToClipboard(text) {
+    chrome.runtime.sendMessage({ type: "copy-to-clipboard", content: text });
+}
+
 export function disableLoginForDomain(domain) {
     chrome.runtime.sendMessage({ type: "disable-login-domain", domain });
 }