+++ webpack.config.js
@@ -45,35 +45,35 @@ module.exports = {
         dialog: path.join(SOURCE, "./dialog/index.js"),
         popup: path.join(SOURCE, "./popup/index.js"),
         setup: path.join(SOURCE, "./setup/index.js"),
-        tab: path.join(SOURCE, "./tab/index.js"),
+        tab: path.join(SOURCE, "./tab/index.js")
     },
 
     module: {
         rules: [
             {
                 test: /\.js$/,
-                use: "babel-loader",
+                use: "babel-loader"
             },
             {
                 test: /\.s[ac]ss$/,
-                use: ["style-loader", "css-loader", "sass-loader"],
+                use: ["style-loader", "css-loader", "sass-loader"]
             },
             {
                 test: /\.css$/,
-                use: ["style-loader", "css-loader"],
+                use: ["style-loader", "css-loader"]
             },
             {
                 test: /\.pug$/,
-                use: "pug-loader",
+                use: "pug-loader"
             },
             {
                 test: /\.(jpg|png|svg|eot|svg|ttf|woff|woff2)$/,
                 loader: "file-loader",
                 options: {
-                    name: "[path][name].[hash].[ext]",
-                },
-            },
-        ],
+                    name: "[path][name].[hash].[ext]"
+                }
+            }
+        ]
     },
 
     node: {
@@ -82,7 +82,7 @@ module.exports = {
         dns: "empty",
         net: "empty",
         stream: "empty",
-        tls: "empty",
+        tls: "empty"
     },
 
     optimization: {
@@ -90,13 +90,13 @@ module.exports = {
             automaticNameDelimiter: "-",
             chunks: "all",
             maxSize: 0,
-            minSize: 30000,
-        },
+            minSize: 30000
+        }
     },
 
     output: {
         filename: "[name].js",
-        path: DIST,
+        path: DIST
     },
 
     plugins: [
@@ -105,49 +105,54 @@ module.exports = {
                 compiler.hooks.afterEmit.tap("AfterEmitPlugin", compilation => {
                     buildManifest(Object.keys(compilation.getStats().compilation.assets));
                 });
-            },
+            }
         },
 
         new CopyWebpackPlugin({
             patterns: [
                 {
-                    from: CHANGELOG,
+                    from: CHANGELOG
                 },
                 {
-                    from: path.join(__dirname, "./resources", "buttercup-*.png"),
+                    from: path.join(__dirname, "./resources", "buttercup-*.png")
                 },
                 {
                     from: ICONS_PATH,
-                    to: "icons",
-                },
-            ],
+                    to: "icons"
+                }
+            ]
         }),
         new DefinePlugin(__configDefines),
         new DefinePlugin({
-            __VERSION__: JSON.stringify(version),
+            __VERSION__: JSON.stringify(version)
         }),
         new HtmlWebpackPlugin({
             title: "Buttercup",
             template: INDEX_TEMPLATE,
             filename: "popup.html",
             inject: "body",
-            chunks: ["popup"],
+            chunks: ["popup"]
         }),
         new HtmlWebpackPlugin({
             title: `Buttercup v${version}`,
             template: INDEX_TEMPLATE,
             filename: "setup.html",
             inject: "body",
-            chunks: ["setup"],
+            chunks: ["setup"]
         }),
         new HtmlWebpackPlugin({
             title: `Buttercup v${version}`,
             template: INDEX_TEMPLATE,
             filename: "dialog.html",
             inject: "body",
-            chunks: ["dialog"],
+            chunks: ["dialog"]
         }),
         new NormalModuleReplacementPlugin(/\/iconv-loader/, "node-noop"),
-        new NormalModuleReplacementPlugin(/random-number-generator|safe-buffer/, "node-noop"),
+        new NormalModuleReplacementPlugin(/random-number-generator|safe-buffer/, "node-noop")
     ],
+
+    watchOptions: {
+        ignored: /node_modules/,
+        poll: 1000
+    }
 };